[JXD-340] Full refactor
