--- esphome/components/automations/automation.cpp
@@ -1,8 +1,0 @@
-#include "esphome/core/log.h"
-#include "esphome/core/helpers.h"
-#include "esphome/core/application.h"
-#include <cstring>
-
-namespace esphome {
-namespace automations {}  // namespace automations
-}  // namespace esphome
