+++ esphome/components/automations/automation_config.h
