+++ esphome/components/automations/enums.cpp
