+++ esphome/components/automations/enums.h
